requirements.txt
Babel==2.6.0
certifi==2019.3.9
Django==2.1.8
django-filter==2.1.0
django-phonenumber-field==2.4.0
django-widget-tweaks==1.4.3
phonenumbers==8.10.11
pytz==2019.1
